Fixed minor bug: RTCP session not fully cleared when it is initialized (should bzero everything), causing incorrect stats when it is reused for next call. Also couple of gcc warnings fixed.
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@569 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjmedia/src/pjmedia/sdp.c
@@ -457,7 +457,7 @@ PJ_DEF(pj_status_t) pjmedia_sdp_rtpmap_to_attr(pj_pool_t *pool,
 			   rtpmap->enc_name.ptr,
 			   rtpmap->clock_rate,
 			   (rtpmap->param.slen ? "/" : ""),
-			   rtpmap->param.slen,
+			   (int)rtpmap->param.slen,
 			   rtpmap->param.ptr);
 
     if (len < 1 || len > sizeof(tempbuf))
